SFR-2676: Deprecate cluster process
