+++ etl-pipeline/tests/functional/processes/ingest/test_loc_process.py
@@ -10,7 +10,7 @@ def test_loc_process(mock_epub_to_webpub):
     number_of_records_ingested = loc_process.runProcess()
 
     record_pipeline_process = RecordPipelineProcess()
-    record_pipeline_process.runProcess(max_attempts=2)
+    record_pipeline_process.runProcess(max_attempts=1)
 
     records = assert_ingested_records(sources=[Source.LOC.value], expected_number_of_records=number_of_records_ingested)
     assert_uploaded_manifests(records)
