+++ etl-pipeline/tests/functional/processes/ingest/test_nypl_process.py
@@ -3,12 +3,13 @@ from processes import NYPLProcess, RecordPipelineProcess
 from .assert_ingested_records import assert_ingested_records
 from .assert_uploaded_manifests import assert_uploaded_manifests
 
+
 def test_nypl_process(mock_epub_to_webpub):
     nypl_process = NYPLProcess('complete', None, None, None, 5, None)
     number_of_records_ingested = nypl_process.runProcess()
 
     record_pipeline_process = RecordPipelineProcess()
-    record_pipeline_process.runProcess(max_attempts=2)
+    record_pipeline_process.runProcess(max_attempts=1)
 
     records = assert_ingested_records(sources=[Source.NYPL.value], expected_number_of_records=number_of_records_ingested)
     assert_uploaded_manifests(records)
