+++ etl-pipeline/tests/functional/processes/ingest/test_publisher_backlist_process.py
@@ -13,7 +13,7 @@ def test_publisher_backlist_process(mock_epub_to_webpub):
     number_of_records_ingested = publisher_backlist_process.runProcess()
 
     record_pipeline_process = RecordPipelineProcess()
-    record_pipeline_process.runProcess(max_attempts=2)
+    record_pipeline_process.runProcess(max_attempts=1)
 
     records = assert_ingested_records(
         sources=[Source.SCHOMBURG.value, Source.U_OF_MICHIGAN_BACKLIST.value],
