+++ etl-pipeline/tests/functional/processes/test_record_pipeline.py
@@ -1,19 +1,10 @@
-import json
-from time import sleep
-
-from managers import RedisManager
 from model import Record
 from processes.record_pipeline import RecordPipelineProcess
-from tests.functional.processes.frbr.test_cluster_process import assert_record_clustered
-from tests.functional.processes.frbr.test_embellish_process import assert_record_embellished
+from .assert_record_clustered import assert_record_clustered
+from .assert_record_embellished import assert_record_embellished
 
 
 def test_record_pipeline(db_manager, unembellished_pipeline_record_uuid, mock_epub_to_webpub):
-    redis_manager = RedisManager()
-
-    redis_manager.create_client()
-    redis_manager.clear_cache()
-
     record_pipeline = RecordPipelineProcess()
 
     record = db_manager.session.query(Record).filter(
@@ -24,9 +15,7 @@ def test_record_pipeline(db_manager, unembellished_pipeline_record_uuid, mock_ep
         message={ "source_id": record.source_id, "source": record.source }
     )
 
-    sleep(1)
-
-    record_pipeline.runProcess(max_attempts=2)
+    record_pipeline.runProcess(max_attempts=1)
 
     assert_record_embellished(record_uuid=unembellished_pipeline_record_uuid, db_manager=db_manager)
     assert_record_clustered(record_uuid=unembellished_pipeline_record_uuid, db_manager=db_manager)
